// prodis-prodiswcl/src/app/modules/prodisapi/model/prospettoProvincia.ts
/*
* SPDX-FileCopyrightText: Copyright 2022 | Regione Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Prodis
 * API per il backend della suite di Prodis.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DatiProvinciali } from './datiProvinciali';
import { Provincia } from './provincia';
import { RiepilogoProvinciale } from './riepilogoProvinciale';


export interface ProspettoProvincia { 
    codUserAggiorn?: string;
    codUserInserim?: string;
    dAggiorn?: Date;
    dInserim?: Date;
    datiProvinciali?: DatiProvinciali;
    flgConfermatoQ2?: string;
    id?: number;
    idProspetto?: number;
    provincia?: Provincia;
    riepilogoProvinciale?: RiepilogoProvinciale;
}